Backend/Code/requirements.txt
cffi==1.14.5
click==8.0.1
Flask==2.0.1
cryptography==3.4.7
Flask-SQLAlchemy==2.5.1
PyMySQL==1.0.2
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
PyJWT==2.1.0
MarkupSafe==2.0.1
pycparser==2.20
SQLAlchemy==1.4.18
Werkzeug==2.0.1